refactor(classifier-host): separate modules into numbered pipeline folders
