+++ addons/classifier-host/src/2-processing/extract-frames.ts
